Lots of changes
Haven't committed in a while...Since the last commit, the menu now works, an app window has been created to hold the frames and there is a new section which holds overdue tasks. Apart from that, there have been lots of other changes to fix up the user experience and make sure that everything works as intended.